Update docs. Consolidate job-runner/main.go (from app/), job/internal/, and router/. Remove unused job/Err* vars.
