+++ .gitignore
@@ -1,0 +1,3 @@
+*.swp
+*.out
+job-runner/job-runner
