+++ job/job.go
@@ -80,13 +80,18 @@ type Factory interface {
 	Make(jobType, jobName string) (Job, error)
 }
 
-// Return represents return values and output from a job. State and Exit
-// indicate how the job completed. Success is STATE_COMPLETE and exit 0.
-// If the job completes with an error, the result is STATE_COMPLETE and a
-// non-zero Exit. Any other State indicates the job did not complete. Error
-// only indicates an internal Go error, not a job error. For example: an RPC
-// client connection error to indicate the job was never attempted because
-// the RPC server couldn't be reached.
+// Return represents return values and output from a job. State indicates how
+// the job completed. If State == proto.STATE_COMPLETE, the job completed
+// successfully. Anything else indicates that the job failed or didn't complete,
+// which will probably cause the job chain to stop. The job is responsible for
+// adhering to this convention.
+//
+// State, Exit, and Error are not mutually exclusive. A job can return State =
+// STATE_COMPLETE to indicate success but also a non-nil Error or a non-zero
+// Exit. This is useful for idempotent jobs and logging that the job was
+// successful because it handled being re-ran. For example, a job could delete
+// a record, but when re-ran the record has already been deleted, so the job
+// is successful but reports Error = ErrRecordNotFound for logging.
 type Return struct {
 	State  byte   // proto/STATE_ const
 	Exit   int64  // Unix exit code
